Got rid of Socket.io!
But... Broke the examples, and bus.client.js for now.

This commit is more of a 'check in' than an interim release.  I've added custom support for WebSocket connections and LongPolling connections, and removed Socket.io from the back end.  We  need to re-write bus.client.js to support the changes, however.   Also, there are plans to support a JSON-P based longpolling solution as well (for cross-domain stuff).